refactor : product image selector
